Mixed ROS1 and ROS2 workspace (#137)
This PR updates the packages in this repo to support building in both a ROS1 and ROS2 environment. The individual packages are not made ROS1 and ROS2 compliant. Instead the ROS2 packages will only build in a ROS2 environment while the ROS1 packages will only build in a ROS1 environment. When in the unsupported environment the packages will log a message stating their compilation has been skipped.